+++ HeatingControl/Application/Commands/SaveHeaterExecutor.cs
@@ -18,7 +18,7 @@ namespace HeatingControl.Application.Commands
         public int PowerOutputDeviceId { get; set; }
         public int PowerOutputChannel { get; set; }
         public UsageUnit UsageUnit { get; set; }
-        public float UsagePerHour { get; set; }
+        public decimal UsagePerHour { get; set; }
         public int MinimumStateChangeIntervalSeconds { get; set; }
     }
